+++ src/Fabulous.AST.Tests/ModuleDeclarations/NestedModule.fs
@@ -55,7 +55,7 @@ module A =
 
     [<Test>]
     let ``Produces a recursive NestedModule`` () =
-        AnonymousModule() { NestedModule("A").toRecursive() { Value("x", "12", false) } }
+        AnonymousModule() { NestedModule("A") { Value("x", "12", false) } |> _.toRecursive() }
 
         |> produces
             """
@@ -67,7 +67,7 @@ module rec A =
 
     [<Test>]
     let ``Produces a private NestedModule`` () =
-        AnonymousModule() { NestedModule("A").accessibility(AccessControl.Private) { Value("x", "12", false) } }
+        AnonymousModule() { NestedModule("A") { Value("x", "12", false) } |> _.toPrivate() }
         |> produces
             """
 
@@ -78,11 +78,77 @@ module private A =
 
     [<Test>]
     let ``Produces a internal NestedModule`` () =
-        AnonymousModule() { NestedModule("A").accessibility(AccessControl.Internal) { Value("x", "12", false) } }
+        AnonymousModule() { NestedModule("A") { Value("x", "12", false) } |> _.toInternal() }
         |> produces
             """
 
 module internal A =
     let x = 12
     
+"""
+
+    [<Test>]
+    let ``Produces a module with nested module`` () =
+        Module("Fabulous.AST") {
+            NestedModule("Foo") {
+                BindingNode(
+                    None,
+                    None,
+                    MultipleTextsNode([ SingleTextNode("let", Range.Zero) ], Range.Zero),
+                    false,
+                    None,
+                    None,
+                    Choice1Of2(IdentListNode([ IdentifierOrDot.Ident(SingleTextNode("x", Range.Zero)) ], Range.Zero)),
+                    None,
+                    List.Empty,
+                    None,
+                    SingleTextNode("=", Range.Zero),
+                    Expr.Constant(Constant.FromText(SingleTextNode("12", Range.Zero))),
+                    Range.Zero
+                )
+            }
+        }
+
+        |> produces
+            """
+module Fabulous.AST
+
+module Foo =
+    let x = 12
+"""
+
+    [<Test>]
+    let ``Produces a module with multiple nested module`` () =
+        Module("Fabulous.AST") {
+            NestedModule("Foo") {
+                BindingNode(
+                    None,
+                    None,
+                    MultipleTextsNode([ SingleTextNode("let", Range.Zero) ], Range.Zero),
+                    false,
+                    None,
+                    None,
+                    Choice1Of2(IdentListNode([ IdentifierOrDot.Ident(SingleTextNode("x", Range.Zero)) ], Range.Zero)),
+                    None,
+                    List.Empty,
+                    None,
+                    SingleTextNode("=", Range.Zero),
+                    Expr.Constant(Constant.FromText(SingleTextNode("12", Range.Zero))),
+                    Range.Zero
+                )
+            }
+
+            NestedModule("Bar") { Value("x", "12", false) }
+        }
+
+        |> produces
+            """
+module Fabulous.AST
+
+module Foo =
+    let x = 12
+
+module Bar =
+    let x = 12
+
 """